Separate Dispatch- types into each header or source files.
