+++ include/eventkit/dispatch/DispatchQueue.h
@@ -9,37 +9,14 @@
 #include <list>
 #include <queue>
 #include <eventkit/common/Allocator.h>
-#include <eventkit/dispatch/RunLoop.h>
+#include <eventkit/common/IntrusivePtr.h>
 #include <eventkit/common/IntrusiveObject.h>
 
 namespace ek {
 namespace dispatch {
 
-class DispatchItem : public ek::common::IntrusiveObject {
-public:
-    explicit DispatchItem(ek::common::Allocator* pA) : ek::common::IntrusiveObject(pA) {}
-
-    virtual ~DispatchItem() = default;
-
-    virtual void run() = 0;
-
-};
-
-template <typename Function>
-class DispatchFunctionItem : public DispatchItem {
-public:
-
-    template <typename F>
-    explicit DispatchFunctionItem(ek::common::Allocator* pA, F&& function);
-
-    ~DispatchFunctionItem() override = default;
-
-    void run() override;
-
-private:
-    Function m_function;
-
-};
+class RunLoop;
+class DispatchItem;
 
 class DispatchQueue : public ek::common::IntrusiveObject {
 public:
